fix: 🐛 Fixes Flameshot file path
The hack I came up with was to create a symlink of my home dir and use
that so that I can have the same save path on all systems. This work
system is the only one with a diff user name and home path so I created
the symlink there.